+++ algo/omegastay/MCTS.h
@@ -28,7 +28,7 @@ namespace MCTS {
     int numVisits = 0;
 
     // Sum of score of all visits from the perspective of the player who just has moved
-    int sumScore = 0;
+    double sumScore = 0;
 
     // Values used for fast calculation of UCT
     double winRate = 0, sqrtLogNumVisits = 0, CDivSqrtNumVisits = 0;
